requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-htmx==0.3.2
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
peewee==3.17.1
psycopg2==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.2
wtf-peewee==3.0.5
WTForms==3.1.2
